src/main/kotlin/com/whammy/article/usecase/ArticleUsecase.kt
package com.whammy.article.usecase

import com.whammy.article.domain.Article
import com.whammy.article.domain.Articles
import com.whammy.article.domain.Comment
import org.springframework.stereotype.Service

@Service
class ArticleUsecase(private val repository: IArticleRepository) {
    fun getArticlesOrderedByUpdatedDateTime(): Articles {
        return repository.getArticles().sortByUpdatedAt()
    }

    fun getArticle(slug: String): Article {
        return repository.getArticle(slug)
    }

    fun getCommentsOfArticle(slug: String): List<Comment> {
        return repository.getCommentsOfArticle(slug)
    }
}
